Fix tests
It turns out that the edge case in 5dfc97198cbc538bfbd03a4bf082e8ebe2fa8e93
is hit exactly twice in the small box: when all particles are active
and the longest active step is longer than the longest particle step,
but the longest step is not a PM step. This only happens in very small
boxes.

The edge case is hit also if the longest active step is longer than the
longest particle step on a PM step, but in this case it does not matter, as
the FullTreeGravAccel has just been updated.